Fixed authorization mistakes

+++ server/routes/orders.js
@@ -12,17 +12,15 @@ router.get("/", async (req, res) => {
 });
 
 router.post("/", async (req, res) => {
-  if (!req.body.user) {
-    return res.status(400).json({ message: "Field user is required!" });
-  }
-
   if (!req.body.dishes) {
     return res.status(400).json({ message: "Field dishes is required!" });
   }
 
   const order = new Order({
-    user: req.body.user,
+    user: req.params.userId,
     dishes: req.body.dishes,
+    paid: false,
+    completed: false,
   });
 
   try {
@@ -35,12 +33,17 @@ router.post("/", async (req, res) => {
 
 router.get("/:orderId", async (req, res) => {
   try {
-    const order = await Order.findById(req.params.orderId).populate("user");
+    const order = await Order.findById(req.params.orderId)
+      .populate("user")
+      .populate("dishes");
     if (!order)
       return res
         .status(404)
         .json({ message: "Order with this ID doesn't exist!" });
 
+    if (order.user.id != req.params.userId) {
+      return res.status(403).json("Forbidden");
+    }
     res.status(200).json(order);
   } catch (err) {
     res.status(500).json({ message: err });
@@ -49,12 +52,15 @@ router.get("/:orderId", async (req, res) => {
 
 router.delete("/:orderId", async (req, res) => {
   try {
-    const order = await Order.findById(req.params.orderId).populate("user");
+    const order = await Order.findById(req.params.orderId);
     if (!order)
       return res
         .status(404)
         .json({ message: "Order with this ID doesn't exist!" });
 
+    if (order.user.id != req.params.userId) {
+      return res.status(403).json("Forbidden");
+    }
     const removedOrder = await Order.deleteOne({
       _id: req.params.orderId,
     });
@@ -65,23 +71,30 @@ router.delete("/:orderId", async (req, res) => {
 });
 
 router.patch("/:orderId", async (req, res) => {
-  if (!req.body.user) {
-    return res.status(400).json({ message: "Field user is required!" });
-  }
-
   if (!req.body.dishes) {
     return res.status(400).json({ message: "Field dishes is required!" });
   }
 
+  if (!req.body.paid) {
+    return res.status(400).json({ message: "Field paid is required!" });
+  }
+
+  if (!req.body.completed) {
+    return res.status(400).json({ message: "Field completed is required!" });
+  }
+
   try {
-    const order = await Order.findById(req.params.orderId).populate("user");
+    const order = await Order.findById(req.params.orderId);
+    if (order.user.id != req.params.userId) {
+      return res.status(403).json("Forbidden");
+    }
     if (!order)
       return res
         .status(404)
         .json({ message: "Order with this ID doesn't exist!" });
 
     const updatedOrder = await Order.findByIdAndUpdate(req.params.orderId, {
-      user: req.body.user,
+      user: req.params.userId,
       dishes: req.body.dishes,
     });
     res.status(200).json(updatedOrder);
